@@ -19,10 +19,8 @@ class PrimaryFieldInfo:
                 if IPrimaryField.providedBy(field):
                     primary = (name, field)
                     break
-        if not primary:
-            raise TypeError("Could not adapt", context, IPrimaryFieldInfo)
-        self.fieldname, self.field = primary
+        self.fieldname, self.field = primary or (None, None)
 
     @property
     def value(self):
-        return self.field.get(self.context)
+        return self.field.get(self.context) if self.field else None
